Likes, work in progress, Need to work on Glyphs

+++ app/controllers/likes_controller.rb
@@ -4,6 +4,7 @@ class LikesController < ApplicationController
   def create
   @bookmark = Bookmark.find(params[:bookmark_id])
   like = current_user.likes.build(bookmark: @bookmark)
+  authorize @bookmark
 
   if like.save
     # Add code to generate a success flash and redirect to @bookmark
@@ -14,12 +15,15 @@ class LikesController < ApplicationController
     flash[:notice] = "Your like was not recorded. Please try again."
     redirect_to @bookmark
   end
+end
 
   def destroy
     # Get the bookmark from the params
     # Find the current user's like with the ID in the param
     @bookmark = Bookmark.find(params[:bookmark_id])
     @like = @user.like(params[:id])
+    authorize @like
+
 
     if like.destroy
       # Flash success and redirect to @bookmark
@@ -30,4 +34,6 @@ class LikesController < ApplicationController
       flash[:notice] = "There was an error. Please try again."
       redirect_to @bookmark
     end
+end
+
 end
